+++ R/calibrate_thresholds.R
@@ -68,9 +68,9 @@ sim_dat1 <- function(p, n) {
 #' Evaluate a single dataset for a single pp_threshold and ppp_threshold
 #' combination
 #'
-#' @description Helper function for calibrate_thresholds() function that evaluates
-#' a single combination of a pp_threshold and a ppp_threshold for a single
-#' dataset
+#' @description Helper function for calibrate_thresholds() function that 
+#' evaluates a single combination of a pp_threshold and a ppp_threshold for a 
+#' single dataset
 #'
 #' @param data the name of the dataset
 #' @param pp_threshold the posterior probability threshold of interest
@@ -86,10 +86,11 @@ sim_dat1 <- function(p, n) {
 #' "greater" if interest is in P(p > p0) and "less" if interest is in P(p < p0).
 #' @param delta clinically meaningful difference between groups.
 #' Typically 0 for the two-sample case. NULL for the one-sample case (default).
+#' @param monitoring the type of interim monitoring to be performed. One of 
+#' "futility" or "efficacy". Default is "futility".
 #' @param prior hyperparameters of prior beta distribution.
 #' Beta(0.5, 0.5) is default
 #' @param S number of samples, default is 5000
-
 #'
 #' @return Returns a tibble with the total sample size at the end of the
 #' trial, the number of responses observed at the end of the trial, the
@@ -101,7 +102,12 @@ sim_dat1 <- function(p, n) {
 #' @importFrom dplyr mutate case_when
 eval_thresh <- function(data, pp_threshold, ppp_threshold, p0, N, 
                         direction = "greater", delta = NULL,
+                        monitoring = "futility",
                         prior = c(0.5, 0.5), S = 5000) {
+  
+  if (!monitoring %in% c("futility", "efficacy"))
+    stop('monitoring must be either "futility" or "efficacy"')
+  
   decision <- NULL
   ppp <- NULL
   for (i in 1:nrow(data)) {
@@ -130,41 +136,26 @@ eval_thresh <- function(data, pp_threshold, ppp_threshold, p0, N,
         theta = pp_threshold
       )
     }
-    decision[i] <- ppp[i] < ppp_threshold
+    decision[i] <- ifelse(monitoring == "futility", 
+                          ppp[i] < ppp_threshold,
+                          ppp[i] > ppp_threshold)
     if (decision[i] == TRUE) break
   }
-  res0 <- add_column(
+  
+  res <- mutate(
     data[ifelse(any(decision == TRUE),
-      which(decision == TRUE),
-      length(decision)
+                which(decision == TRUE),
+                length(decision)
     ), ],
     pp_threshold = pp_threshold,
     ppp_threshold = ppp_threshold,
     ppp = ppp[ifelse(any(decision == TRUE),
-      which(decision == TRUE),
-      length(decision)
-    )]
+                     which(decision == TRUE),
+                     length(decision)
+    )],
+    positive = ppp > pp_threshold
   )
-
-  if (ncol(data) == 4) {
-    res <- mutate(
-      res0,
-      positive = case_when(
-        sum(n0, n1) == sum(N) & ppp > pp_threshold ~ TRUE,
-        sum(n0, n1) == sum(N) & ppp <= pp_threshold ~ FALSE,
-        sum(n0, n1) != sum(N) ~ FALSE
-      )
-    )
-  } else if (ncol(data) == 2) {
-    res <- mutate(
-      res0,
-      positive = case_when(
-        n1 == N & ppp > pp_threshold ~ TRUE,
-        n1 == N & ppp <= pp_threshold ~ FALSE,
-        n1 != N ~ FALSE
-      )
-    )
-  }
+  
   return(res)
 }
 
@@ -175,12 +166,11 @@ eval_thresh <- function(data, pp_threshold, ppp_threshold, p0, N,
 #' probability threshold with interim futility monitoring
 #'
 #' @description This function is meant to be used in the context of a
-#' clinical trial with a binary endpoint. For a vector of possible posterior
-#' decision thresholds, the function simulates many trials and then calculates
-#' the average number of times the posterior probability exceeds a given
-#' threshold. In a null case, this will result in the type I error at a given
-#' threshold. In an alternative case, this will result in the power at a given
-#' threshold.
+#' clinical trial with a binary endpoint. For every combination of the provided
+#' posterior thresholds and predictive thresholds, the function simulates many 
+#' trials and then calculates the average number of times a trial was positive. 
+#' In the null case, this is the type I error for the given thresholds. 
+#' In the alternative case, this is the power for the given thresholds.
 #'
 #' @param p_null vector of length two containing the probability of event in
 #' the standard of care and experimental arm c(p0, p1) for the two-sample case
@@ -203,13 +193,15 @@ eval_thresh <- function(data, pp_threshold, ppp_threshold, p0, N,
 #' for two-sample case; integer of total planned sample size at end of trial N
 #' for one-sample case
 #' @param pp_threshold the posterior probability threshold of interest
-#' @param ppp_threshold the posterior probability threshold of interest for
-#' futility monitoring
+#' @param ppp_threshold the posterior predictive probability threshold of 
+#' interest for futility monitoring
 #' @param direction "greater" (default) if interest is in p(p1 > p0) and "less"
 #' if interest is in p(p1 < p0) for two-sample case. For one-sample case,
 #' "greater" if interest is in p(p > p0) and "less" if interest is in p(p < p0).
 #' @param delta clinically meaningful difference between groups.
 #' Typically 0 for the two-sample case. NULL for the one-sample case (default).
+#' @param monitoring the type of interim monitoring to be performed. One of 
+#' "futility" or "efficacy". Default is "futility".
 #' @param prior hyperparameters of prior beta distribution.
 #' Beta(0.5, 0.5) is default
 #' @param S number of samples drawn from the posterior. Default is 5000
@@ -235,34 +227,36 @@ eval_thresh <- function(data, pp_threshold, ppp_threshold, p0, N,
 #' for a null setting, and a measure of the power in the alternative setting.
 #'
 #' @examples
-#'
+#' 
 #' # One-sample case
-#' \donttest{
 #' set.seed(123)
-#'
+#' 
 #' calibrate_thresholds(
-#'   p_null = 0.1, p_alt = 0.3,
-#'   n = seq(5, 25, 5), N = 25, 
-#'   pp_threshold = c(0.9, 0.95, 0.96, 0.98),
-#'   ppp_threshold = seq(0.05, 0.2, 0.05),
-#'   direction = "greater", delta = NULL,
-#'   prior = c(0.5, 0.5), S = 5000, nsim = 1000
-#' )
-#' }
-#'
+#'   p_null = 0.1, 
+#'   p_alt = 0.4,
+#'   n = seq(5, 15, 5), 
+#'   N = 15,
+#'   pp_threshold = c(0.85, 0.9),
+#'   ppp_threshold = c(0.1, 0.2),
+#'   S = 10, 
+#'   nsim = 10
+#'   )
+#' 
 #' # Two-sample case
-#' \donttest{
-#' set.seed(123)
-#'
+#' set.seed(456)
+#' 
 #' calibrate_thresholds(
-#'   p_null = c(0.1, 0.1), p_alt = c(0.1, 0.3),
-#'   n = cbind(seq(5, 25, 5), seq(5, 25, 5)), N = c(25, 25), 
-#'   pp_threshold = c(0.9, 0.95, 0.96, 0.98),
-#'   ppp_threshold = seq(0.05, 0.2, 0.05),
-#'   direction = "greater", delta = 0,
-#'   prior = c(0.5, 0.5), S = 5000, nsim = 1000
-#' )
-#' }
+#'   p_null = c(0.1, 0.1), 
+#'   p_alt = c(0.1, 0.5),
+#'   n = cbind(seq(5, 15, 5), seq(5, 15, 5)), 
+#'   N = c(15, 15),
+#'   pp_threshold = c(0.8, 0.85),
+#'   ppp_threshold = c(0.2, 0.3),
+#'   delta = 0,
+#'   S = 10, 
+#'   nsim = 10
+#'   )
+#' 
 #' @importFrom purrr pmap_dfr map cross_df
 #' @importFrom furrr future_map furrr_options
 #' @importFrom dplyr bind_rows full_join select rename ungroup summarize
@@ -271,20 +265,21 @@ eval_thresh <- function(data, pp_threshold, ppp_threshold, p0, N,
 calibrate_thresholds <- function(p_null, p_alt, n, N,
                                  pp_threshold, ppp_threshold, 
                                  direction = "greater", delta = NULL, 
+                                 monitoring = "futility",
                                  prior = c(0.5, 0.5),
                                  S = 5000, nsim = 1000) {
   sim_dat_null <-
     map(seq_len(nsim), ~ sim_dat1(p = p_null, n = n))
-
+  
   sim_dat_alt <-
     map(seq_len(nsim), ~ sim_dat1(p = p_alt, n = n))
-
+  
   cross_threshold <-
     cross_df(list(
       pp_threshold = pp_threshold,
       ppp_threshold = ppp_threshold
     ))
-
+  
   p0 <- if(length(p_null) == 1) 
     p_null else if(length(p_null) == 2)
       NULL
@@ -295,10 +290,16 @@ calibrate_thresholds <- function(p_null, p_alt, n, N,
       function(x) 
         pmap_dfr(cross_threshold,
                  function(pp_threshold, ppp_threshold) 
-                   eval_thresh(x, pp_threshold, ppp_threshold,
-                 direction = direction, p0 = p0, 
-                 delta = delta, prior = prior, 
-                 S = S, N = N)), 
+                   eval_thresh(x, 
+                               pp_threshold, 
+                               ppp_threshold,
+                               direction = direction, 
+                               p0 = p0, 
+                               delta = delta, 
+                               prior = prior, 
+                               S = S, 
+                               N = N,
+                               monitoring = monitoring)), 
       .options = furrr_options(seed = TRUE)
     )
   
@@ -308,46 +309,52 @@ calibrate_thresholds <- function(p_null, p_alt, n, N,
       function(x) 
         pmap_dfr(cross_threshold,
                  function(pp_threshold, ppp_threshold) 
-                   eval_thresh(x, pp_threshold, ppp_threshold,
-                 direction = direction, p0 = p0, 
-                 delta = delta, prior = prior, 
-                 S = S, N = N)),
+                   eval_thresh(x, 
+                               pp_threshold, 
+                               ppp_threshold,
+                               direction = direction, 
+                               p0 = p0, 
+                               delta = delta, 
+                               prior = prior, 
+                               S = S, 
+                               N = N,
+                               monitoring = monitoring)),
       .options = furrr_options(seed = TRUE)
     )
-
+  
   res_df_null <-
     bind_rows(
       res_null,
       .id = "sim_num"
     )
-
+  
   res_df_alt <-
     bind_rows(
       res_alt,
       .id = "sim_num"
     )
-
+  
   if (length(p_null) == 2) {
     res_df <-
       full_join(
         select(
           rename(res_df_null,
-            n0_null = n0, 
-            n1_null = n1,
-            positive_null = positive
+                 n0_null = n0, 
+                 n1_null = n1,
+                 positive_null = positive
           ),
           -ppp, -y0, -y1
         ),
         select(
           rename(res_df_alt,
-            n0_alt = n0, 
-            n1_alt = n1,
-            positive_alt = positive
+                 n0_alt = n0, 
+                 n1_alt = n1,
+                 positive_alt = positive
           ),
           -ppp, -y0, -y1
         )
       )
-
+    
     res_summary <-
       ungroup(
         summarize(
@@ -374,7 +381,7 @@ calibrate_thresholds <- function(p_null, p_alt, n, N,
           -ppp, -y1
         )
       )
-
+    
     res_summary <-
       ungroup(
         summarize(
@@ -388,13 +395,13 @@ calibrate_thresholds <- function(p_null, p_alt, n, N,
         )
       )
   }
-
+  
   # create empty list to return everything
   x <- list()
-
+  
   # add the main results in
   x$res_summary <- res_summary
-
+  
   # will return call, and all object passed to in table1 call
   # the object func_inputs is a list of every object passed to the function
   calibrate_thresholds_inputs <- list(
@@ -408,9 +415,10 @@ calibrate_thresholds <- function(p_null, p_alt, n, N,
     N = N,
     nsim = nsim,
     pp_threshold = pp_threshold,
-    ppp_threshold = ppp_threshold
+    ppp_threshold = ppp_threshold,
+    monitoring = monitoring
   )
-
+  
   # create other objects to return
   x$call_list <- list(calibrate_thresholds = match.call())
   x$inputs <- calibrate_thresholds_inputs
@@ -421,4 +429,3 @@ calibrate_thresholds <- function(p_null, p_alt, n, N,
   x
   
 }
-