+++ inst/doc/one_sample_expansion.R
@@ -1,27 +1,35 @@
 ## ---- include = FALSE---------------------------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE,
-  comment = "#>"
+  comment = "#>",
+  warning = FALSE, 
+  message = FALSE, 
+  error = FALSE
 )
 
-## ----setup, warning = FALSE, message = FALSE, error = FALSE-------------------
-# remotes::install_github("zabore/ppseq")
+## ----eval = FALSE-------------------------------------------------------------
+#  install.packages("ppseq")
+#  install.packages("future")
+
+## -----------------------------------------------------------------------------
 library(ppseq)
-library(future)
+
+## ----eval = FALSE-------------------------------------------------------------
+#  library(future)
 
 ## ----eval = FALSE-------------------------------------------------------------
 #  set.seed(123)
 #  
-#  future::plan(future::multicore(workers = 76))
+#  future::plan(future::multicore(workers = 56))
 #  
 #  one_sample_cal_tbl <-
 #    calibrate_thresholds(p_null = 0.1,
 #                         p_alt = 0.2,
 #                         n = seq(5, 95, 5),
 #                         N = 95,
-#                         pp_threshold = c(0, 0.7, 0.74, 0.78, 0.82, 0.86, 0.9,
+#                         pp_threshold = c(0.7, 0.74, 0.78, 0.82, 0.86, 0.9,
 #                                          0.92, 0.93, 0.94, 0.95, 0.96, 0.97,
-#                                          0.98, 0.99, 0.999, 0.9999, 0.99999, 1),
+#                                          0.98, 0.99),
 #                         ppp_threshold = seq(0.05, 0.2, 0.05),
 #                         direction = "greater",
 #                         delta = NULL,
@@ -61,6 +69,9 @@ optimize_design(one_sample_cal_tbl,
 #      prior = c(0.5, 0.5),
 #      S = 5000
 #      )
+
+## ----eval = FALSE-------------------------------------------------------------
+#  one_sample_decision_tbl
 
 ## ----echo = FALSE-------------------------------------------------------------
 gt::gt(one_sample_decision_tbl)