+++ NEWS.md
@@ -1,3 +1,18 @@
+# ppseq 0.2.0
+
+* added weighting options to `optimize_design()` (Closed Issue #7)
+* added informative messaging when no results returned from `optimize_design()` (Closed Issue #14)
+* upgraded package website to bootstrap 5 and changed some styling
+* added two-sample randomized trial vignette
+* added `calc_next()` function to calculate probability of response in next patient (Closed Issue #2)
+* fixed a bug in the default arguments to `optimize_design()` (Closed Issue #12)
+* added efficacy monitoring option, default is still futility monitoring (Closed Issue #3)
+
+# ppseq 0.1.2
+
+* Updating all examples to run without `\donttest{}` 
+* Some minor documentation changes
+
 # ppseq 0.1.1
 
 * Updating the included example data files
@@ -7,8 +22,3 @@
 # ppseq 0.1.0
 
 * Initial package release
-
-
-# ppseq 0.0.0.9000
-
-* Added a `NEWS.md` file to track changes to the package.